consolidated server-creation into 1 utils module; added unit tests
