+++ libraries/botframework-streaming/src/index.ts
@@ -12,6 +12,7 @@ export {
     INodeBuffer,
     INodeIncomingMessage,
     INodeSocket,
+    INodeSocket2,
     IReceiveRequest,
     IReceiveResponse,
     ISocket,
